+++ sentry_sdk/integrations/openai.py
@@ -2,9 +2,9 @@ from functools import wraps
 
 from sentry_sdk import consts
 from sentry_sdk._types import TYPE_CHECKING
-from sentry_sdk.ai_monitoring import record_token_usage
+from sentry_sdk.ai.monitoring import record_token_usage
 from sentry_sdk.consts import SPANDATA
-from sentry_sdk.integrations._ai_common import set_data_normalized
+from sentry_sdk.ai.utils import set_data_normalized
 
 if TYPE_CHECKING:
     from typing import Any, Iterable, List, Optional, Callable, Iterator
